external/webkit/LayoutTests/platform/android/webgl-conformance/dumpAsText/conformance/rendering/triangle-expected.txt

Rendering Test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS drawing is correct
PASS successfullyParsed is true

TEST COMPLETE
